create password form component, sign up modal refactoring
